Form gender options and underline change
Gender option is now implemented via images instead of radio buttons. UI changes
are done via jQuery. The reason for this is because it was too difficult to
override Materialize default behaviour. There was one mention on how to
override on the net. However, upon testing, this method did not pan out.

Updated label underline focus color to purple.