Transitions for form body, sidebar, and lines are in place.
Initially, the transitions were driven by the update property for each movement.
This approach did not work out, as it triggers right away without waiting for
the delay. After exploring the approach of using the begin method, it turned out
that there was no reliable way to progress percentage. This meant reverting back
to the update method and manually calculating the progress. The issue with this
approach, is that you cannot use any easing function. We must rely on a
percentage instead of width, as the clip-path is using such values.